pyproject.toml
[project]
name = "example-python-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.12.20",
    "uvicorn>=0.34.3",
]
